airflow/requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-google
boto3
google-cloud-bigquery
stripe
confluent-kafka
python-dotenv